Rename the gcp module to publisher
